reafactor: adding chack id and auth.id

+++ src/domain/users/useCase/DeleteUserUseCase.js
@@ -3,6 +3,7 @@ const { userRepository } = require("../../repository/index");
 class DeleteUserUseCase {
     async findOne(data) {
         const userId = data.params.id;
+        const authId = data.auth;
 
         const userDel = await userRepository.findOne(userId);
 
@@ -10,13 +11,13 @@ class DeleteUserUseCase {
             return new Error("Nenhum usuario cadastrado com este ID!");
         }
 
-        const userToDelete = await userRepository.destroy(userId);
-
-        if (!userToDelete) {
-            return new Error("Nenhum usuario cadastrado com este ID!");
+        if (!userId != authId) {
+            return new Error("Você não pode realizar essa ação!");
         }
 
-        return userToDelete;
+        await userRepository.destroy(userId);
+
+        return;
     }
 }
